Refactor UserSavedModal to standalone component
fixes misc bugs with enter button opening paper during list creation
